benchmark: move the benchmark suite out of tests/ into benchmarks/
phase1_prompt_benchmarking is a benchmarking pipeline, not a test
suite; it moves wholesale to benchmarks/ (pure rename -- the
benchmarking_outputs submodule path, import lines, path constants, and
the ruff per-file ignore follow). The unit tests that exercise it stay
in tests/ and import from the new location.
