+++ benchmarks/phase1_prompt_benchmarking/benchmark_bundles/aconcagua_interphase__cluster_41__bundle.json
