+++ benchmarks/phase1_prompt_benchmarking/benchmark_bundles/aconcagua_interphase_shuffled__cluster_17__bundle.json
