+++ benchmarks/phase1_prompt_benchmarking/benchmark_bundles/denali__cluster_24__bundle.json
