+++ benchmarks/phase1_prompt_benchmarking/benchmark_bundles/denali__cluster_43__bundle.json
