+++ benchmarks/phase1_prompt_benchmarking/benchmark_bundles/jebel__cluster_0__bundle.json
