+++ benchmarks/phase1_prompt_benchmarking/benchmark_bundles/whitney__cluster_49__bundle.json
