+++ benchmarks/phase1_prompt_benchmarking/benchmark_clusters_ground_truth.csv
