+++ benchmarks/phase1_prompt_benchmarking/benchmark_inputs/aconcagua_interphase_screen_context.json
