+++ benchmarks/phase1_prompt_benchmarking/benchmark_inputs/denali_screen_context.json
