+++ benchmarks/phase1_prompt_benchmarking/benchmark_inputs/whitney_screen_context.json
