+++ tests/test_bench_dry_run.py
@@ -11,12 +11,12 @@ _REPO_ROOT = Path(__file__).resolve().parents[1]
 if str(_REPO_ROOT) not in sys.path:
     sys.path.insert(0, str(_REPO_ROOT))
 
-from tests.phase1_prompt_benchmarking.architecture_benchmarking_workflow.bench_dry_run import (
+from benchmarks.phase1_prompt_benchmarking.architecture_benchmarking_workflow.bench_dry_run import (
     _load_bundle_genes,
     generate_mock_parsed_output,
     generate_mock_raw_outputs,
 )
-from tests.phase1_prompt_benchmarking.architecture_benchmarking_workflow.bench_routes import (
+from benchmarks.phase1_prompt_benchmarking.architecture_benchmarking_workflow.bench_routes import (
     ROUTE_REGISTRY,
 )
