+++ tests/test_order_bench_orderings.py
@@ -10,12 +10,11 @@ _REPO_ROOT = Path(__file__).resolve().parents[1]
 if str(_REPO_ROOT) not in sys.path:
     sys.path.insert(0, str(_REPO_ROOT))
 
-from mozzarellm.prompt_components import COMPONENT_REGISTRY
-from tests.phase1_prompt_benchmarking.architecture_benchmarking_workflow.bench_routes import (
+from benchmarks.phase1_prompt_benchmarking.architecture_benchmarking_workflow.bench_routes import (
     ROUTE_REGISTRY,
     StepwiseTurn,
 )
-from tests.phase1_prompt_benchmarking.architecture_benchmarking_workflow.order_bench_orderings import (
+from benchmarks.phase1_prompt_benchmarking.architecture_benchmarking_workflow.order_bench_orderings import (
     _ORDER_SPECS,
     ORDER_VARIANTS,
     OrderVariant,
@@ -23,6 +22,7 @@ from tests.phase1_prompt_benchmarking.architecture_benchmarking_workflow.order_b
     resolve_order_variant_ids,
     validate_order_variant_names,
 )
+from mozzarellm.prompt_components import COMPONENT_REGISTRY
 
 # ============================================================================
 # Variant registry tests
